# dist/utils/swagger-output.json
{
    "swagger": "2.0",
    "info": {
        "title": "Quizmon API",
        "description": "REST API documentation for the Quizmon quiz platform. Supports quiz creation, match hosting, AI-powered question generation, and real-time gameplay via Socket.IO.",
        "version": "1.0.0"
    },
    "host": "",
    "basePath": "/",
    "schemes": [
        "http"
    ],
...
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization",
            "description": "JWT Bearer token. Format: `Bearer <token>`"
        }
    },
    "paths": {
...
        "/question/{id}": {
            "get": {
                "tags": [
                    "Question"
                ],
                "summary": "Get a single question by ID",
                "description": "Retrieves a question and all its type-specific data by question ID.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 10
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question returned"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/question/buttons": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Create a multiple-choice (buttons) question",
                "description": "Creates an MCQ question with selectable button options. Accepts optional media file uploads.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "example": 1
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "example": "What is the capital of France?"
                    },
                    {
                        "name": "options",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": "JSON array of options: [{text, isCorrect, order}]"
                    },
                    {
                        "name": "files",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Question created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/buttons/{id}": {
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Update a buttons question",
                "description": "Updates an existing multiple-choice (buttons) question. Accepts optional media file uploads.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 5
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "options",
                        "in": "formData",
                        "required": false,
                        "type": "string",
                        "description": "JSON array of options"
                    },
                    {
                        "name": "files",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/checkboxes": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Create a multi-select (checkboxes) question",
                "description": "Creates a question where users can select multiple correct answers.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "example": 1
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "example": "Which of these are planets?"
                    },
                    {
                        "name": "options",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": "JSON array of options: [{text, isCorrect, order}]"
                    },
                    {
                        "name": "files",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Question created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/checkboxes/{id}": {
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Update a checkboxes question",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 5
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "options",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "files",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/question/reorder": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Create a drag-to-reorder question",
                "description": "Creates a question where users must arrange options in the correct order.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "example": 1
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "example": "Order the planets by distance from the Sun"
                    },
                    {
                        "name": "options",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "description": "JSON array of options with correct order: [{text, order}]"
                    },
                    {
                        "name": "files",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Question created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/reorder/{id}": {
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Update a reorder question",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 5
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "options",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "files",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/location": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Create a map location question",
                "description": "Creates a question where users must pin a location on a map. The answer is validated by proximity to the correct coordinates.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "quizId",
                                "text",
                                "correctLatitude",
                                "correctLongitude"
                            ],
                            "properties": {
                                "quizId": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "text": {
                                    "type": "string",
                                    "example": "Where is the Eiffel Tower?"
                                },
                                "correctLatitude": {
                                    "type": "number",
                                    "example": 48.8584
                                },
                                "correctLongitude": {
                                    "type": "number",
                                    "example": 2.2945
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Question created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/location/{id}": {
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Update a location question",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 5
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "text": {
                                    "type": "string"
                                },
                                "correctLatitude": {
                                    "type": "number"
                                },
                                "correctLongitude": {
                                    "type": "number"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/typeanswer": {
            "post": {
                "tags": [
                    "Question"
                ],
                "summary": "Create a type-answer (text input) question",
                "description": "Creates a question where users must type a text answer that is matched against the correct answer.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "quizId",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "example": 1
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "example": "What is the chemical symbol for gold?"
                    },
                    {
                        "name": "correctAnswer",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "example": "Au"
                    },
                    {
                        "name": "files",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Question created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/question/typeanswer/{id}": {
            "put": {
                "tags": [
                    "Question"
                ],
                "summary": "Update a type-answer question",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 5
                    },
                    {
                        "name": "text",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "correctAnswer",
                        "in": "formData",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "files",
                        "in": "formData",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/ai/jobs/{id}/questions/{questionId}": {
            "put": {
                "tags": [
                    "AI Generation"
                ],
                "summary": "Review a generated question (approve / reject / regenerate)",
                "description": "Updates the status of a generated question. Status values: `PENDING`, `APPROVED`, `REJECTED`, `REGENERATING`. Optionally include feedback for regeneration.",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 3
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "example": 18
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "status"
                            ],
                            "properties": {
                                "status": {
                                    "type": "string",
                                    "enum": [
                                        "PENDING",
                                        "APPROVED",
                                        "REJECTED",
                                        "REGENERATING"
                                    ],
                                    "example": "APPROVED"
                                },
                                "userFeedback": {
                                    "type": "string",
                                    "example": "Make this question harder"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Question updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Job not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    }
}